If --x86-report and --min-inst-per-ctx print report
Print a report of the x86 cpu every time a ctx reaches the minimum of
instructions. The report is stored next to the normal --x86-report
printed at the ent of the execution but its format is "<x86-report>.pid<number>".
Also, the first core and thread that appear in the report are the ones
this ctx was last mapped.